requirement.txt
streamlit==1.26.0
stegano==0.9.3
Pillow==10.0.0